refactor: export download functions into file
