+++ web-ui/src/utils/download.js
@@ -1,0 +1,15 @@
+export function getFilename(contentDisposition) {
+  const match = contentDisposition?.match(/filename="?([^";]+)"?/i);
+  return match?.[1]?.trim() ?? null;
+}
+
+export function triggerBlobDownload(blob, filename) {
+  const blobUrl = window.URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = blobUrl;
+  a.download = filename;
+  document.body.appendChild(a);
+  a.click();
+  a.remove();
+  window.URL.revokeObjectURL(blobUrl);
+}
